move variables to shared folder
